Platforms/Armada80x0: add ACPI description of the PCI root
Add a MCFG table and add a description to the DSDT of the PCIe root
complex on the 0x70x0/80x0 SOCs. Note that the Synopsys IP is not
fully compliant, and requires a workaround in order to support ACPI
boot which reduces the bus range to a single bus.

Contributed-under: TianoCore Contribution Agreement 1.0
Signed-off-by: Ard Biesheuvel <ard.biesheuvel@linaro.org>